requirements.txt
bugjar==0.1.0
Django==1.7
django-appconf==1.0.1
django-compressor==1.5
django-debug-toolbar==1.3.0
django-extensions==1.5.5
django-image-cropping==1.0.2
django-localflavor==1.1
easy-thumbnails==2.2
factory-boy==2.5.2
fake-factory==0.5.3
funcsigs==0.4
mock==1.3.0
model-mommy==1.2.5
pbr==1.8.0
pdb==0.1
Pillow==2.7.0
Pygments==2.0.2
python-gnupg==0.3.7
pytz==2015.4
PyYAML==3.11
six==1.9.0
sqlparse==0.1.15
tkreadonly==0.6.0
Werkzeug==0.10.4
wheel==0.24.0
